fix(rich-text): preserve literal tilde serialization
